backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
# Database - PostgreSQL (Neon)
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
alembic>=1.13.0
# Database - MongoDB (for migration only)
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
typer>=0.9.0
resend>=0.7.0
redis>=5.0.0
sentry-sdk[fastapi]==2.0.0
slowapi>=0.1.9

httpx>=0.27.0
pytest-asyncio>=0.23.0
